AUDIT-54:Added test cases for ModuleEventDao and tweaked return type to long for count methods

+++ api/src/main/java/org/openmrs/module/auditlogweb/api/ModuleEventService.java
@@ -25,5 +25,5 @@ public interface ModuleEventService {
 	
 	List<ModuleEvent> getRelatedModuleEvents(String sessionId, int page, int size);
 	
-	Integer countRelatedModuleEvents(String sessionId);
+	long countRelatedModuleEvents(String sessionId);
 }
